test: extract waitForState helper
