+++ tests/helpers/classicBattle/battleStateBadge.test.js
@@ -13,25 +13,7 @@ import {
   mockRoundSelectModal
 } from "./mocks.js";
 import { CLASSIC_BATTLE_STATES } from "../../../src/helpers/classicBattle/stateTable.js";
-import { onBattleEvent, offBattleEvent } from "../../../src/helpers/classicBattle/battleEvents.js";
-import { getStateSnapshot } from "../../../src/helpers/classicBattle/battleDebug.js";
-
-function waitForState(target, timeout = 10000) {
-  return new Promise((resolve, reject) => {
-    if (getStateSnapshot().state === target) return resolve();
-    const handler = (e) => {
-      if (e.detail?.to === target) {
-        offBattleEvent("battleStateChange", handler);
-        resolve();
-      }
-    };
-    onBattleEvent("battleStateChange", handler);
-    setTimeout(() => {
-      offBattleEvent("battleStateChange", handler);
-      reject(new Error(`timeout for ${target}`));
-    }, timeout);
-  });
-}
+import { waitForState } from "../../waitForState.js";
 
 // Apply all the necessary mocks
 mockScheduler();
