+++ tests/helpers/classicBattle/scheduleNextRound.test.js
@@ -3,29 +3,8 @@ import "./commonMocks.js";
 import { setupClassicBattleDom } from "./utils.js";
 import { createTimerNodes } from "./domUtils.js";
 import { applyMockSetup } from "./mockSetup.js";
-import {
-  onBattleEvent,
-  offBattleEvent,
-  emitBattleEvent
-} from "../../../src/helpers/classicBattle/battleEvents.js";
-import { getStateSnapshot } from "../../../src/helpers/classicBattle/battleDebug.js";
-
-function waitForState(target, timeout = 10000) {
-  return new Promise((resolve, reject) => {
-    if (getStateSnapshot().state === target) return resolve();
-    const handler = (e) => {
-      if (e.detail?.to === target) {
-        offBattleEvent("battleStateChange", handler);
-        resolve();
-      }
-    };
-    onBattleEvent("battleStateChange", handler);
-    setTimeout(() => {
-      offBattleEvent("battleStateChange", handler);
-      reject(new Error(`timeout for ${target}`));
-    }, timeout);
-  });
-}
+import { emitBattleEvent } from "../../../src/helpers/classicBattle/battleEvents.js";
+import { waitForState } from "../../waitForState.js";
 
 vi.mock("../../../src/helpers/CooldownRenderer.js", () => ({
   attachCooldownRenderer: vi.fn()
